XP-679 Publish wiz show children and dependencies - [Modified] name and behavior of ShowDependencies label - [Modified] dialog to show chidlren in the same list as dependents - [Modified] naming of some of the classes used for clearer distinction of usage purposes
