+++ modules/admin-impl/src/test/java/com/enonic/xp/admin/impl/rest/resource/content/ResolvePublishRequestedContentsResultJsonFactoryTest.java
@@ -4,8 +4,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 
-import com.enonic.xp.admin.impl.rest.resource.content.json.ResolveDependantsResultJson;
-import com.enonic.xp.admin.impl.rest.resource.content.json.ResolvePublishDependenciesResultJson;
+import com.enonic.xp.admin.impl.rest.resource.content.json.ResolvePublishRequestedContentsResultJson;
 import com.enonic.xp.content.CompareContentResult;
 import com.enonic.xp.content.CompareContentResults;
 import com.enonic.xp.content.CompareStatus;
@@ -18,11 +17,15 @@ import com.enonic.xp.content.ResolvePublishDependenciesResult;
 import com.enonic.xp.schema.content.ContentTypeName;
 import com.enonic.xp.schema.content.ContentTypeService;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
 
-public class ResolveDependantsResultJsonFactoryTest
+public class ResolvePublishRequestedContentsResultJsonFactoryTest
 {
+
     private ContentTypeService contentTypeService;
+
 
     @Before
     public void before()
@@ -48,9 +51,20 @@ public class ResolveDependantsResultJsonFactoryTest
         final ResolvePublishDependenciesResult resolvedDependencies =
             ResolvePublishDependenciesResult.create().pushContentRequests( pushRequests ).build();
 
-        final ResolveDependantsResultJson result = createDependantsResultJson( resolvedDependencies, resolved, compareResults );
+        final ResolvePublishRequestedContentsResultJson result =
+            createPublishDependenciesResultJson( resolvedDependencies, resolved, compareResults );
 
-        assertEquals( 0, result.getDependantContents().size() );
+        assertEquals( 1, result.getPushRequestedContents().size() );
+        assertEquals( 2, result.getPushRequestedContents().get( 0 ).getChildrenCount() );
+        assertEquals( 0, result.getPushRequestedContents().get( 0 ).getDependantsCount() );
+
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getCompareStatus() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getName() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getDisplayName() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getIconUrl() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getId() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getPath() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getType() );
     }
 
     @Test
@@ -70,57 +84,20 @@ public class ResolveDependantsResultJsonFactoryTest
         final ResolvePublishDependenciesResult resolvedDependencies =
             ResolvePublishDependenciesResult.create().pushContentRequests( pushRequests ).build();
 
-        final ResolveDependantsResultJson result = createDependantsResultJson( resolvedDependencies, resolved, compareResults );
+        final ResolvePublishRequestedContentsResultJson result =
+            createPublishDependenciesResultJson( resolvedDependencies, resolved, compareResults );
 
-        assertEquals( 2, result.getDependantContents().size() );
+        assertEquals( 1, result.getPushRequestedContents().size() );
+        assertEquals( 0, result.getPushRequestedContents().get( 0 ).getChildrenCount() );
+        assertEquals( 2, result.getPushRequestedContents().get( 0 ).getDependantsCount() );
 
-        assertNotNull( result.getDependantContents().get( 0 ).getCompareStatus() );
-        assertNotNull( result.getDependantContents().get( 0 ).getName() );
-        assertNotNull( result.getDependantContents().get( 0 ).getDisplayName() );
-        assertNotNull( result.getDependantContents().get( 0 ).getIconUrl() );
-        assertNotNull( result.getDependantContents().get( 0 ).getId() );
-        assertNotNull( result.getDependantContents().get( 0 ).getPath() );
-        assertNotNull( result.getDependantContents().get( 0 ).getType() );
-    }
-
-    @Test
-    public void testGeneratedJsonOrderedByPath()
-    {
-
-        final Content content1 = createContent( "s1", "s1Name", ContentPath.from( "b" ), true );
-        final Content content2 = createContent( "s2", "s1Name", ContentPath.from( "a" ), true );
-        final Content content3 = createContent( "s3", "s1Name", ContentPath.from( "c" ), true );
-        final Content content4 = createContent( "s4", "s1Name", ContentPath.from( "s4" ), true );
-
-        final Contents resolved = Contents.from( content1, content2, content3, content4 );
-
-        final CompareContentResults compareResults = createCompareContentResults();
-
-        final PushContentRequests pushRequests = PushContentRequests.create().
-            addParentOf( ContentId.from( "s1" ), ContentId.from( "s2" ), ContentId.from( "s4" ) ).
-            addParentOf( ContentId.from( "s2" ), ContentId.from( "s3" ), ContentId.from( "s4" ) ).
-            addParentOf( ContentId.from( "s3" ), ContentId.from( "s4" ), ContentId.from( "s4" ) ).
-            addRequested( ContentId.from( "s4" ), ContentId.from( "s4" ) ).
-            build();
-
-        final ResolvePublishDependenciesResult resolvedDependencies =
-            ResolvePublishDependenciesResult.create().pushContentRequests( pushRequests ).build();
-
-        final ResolveDependantsResultJson result = createDependantsResultJson( resolvedDependencies, resolved, compareResults );
-
-        assertEquals( 3, result.getDependantContents().size() );
-
-        assertEquals( result.getDependantContents().get( 0 ).getId(), "s2" );
-        assertEquals( result.getDependantContents().get( 1 ).getId(), "s1" );
-        assertEquals( result.getDependantContents().get( 2 ).getId(), "s3" );
-
-        assertNotNull( result.getDependantContents().get( 0 ).getCompareStatus() );
-        assertNotNull( result.getDependantContents().get( 0 ).getName() );
-        assertNotNull( result.getDependantContents().get( 0 ).getDisplayName() );
-        assertNotNull( result.getDependantContents().get( 0 ).getIconUrl() );
-        assertNotNull( result.getDependantContents().get( 0 ).getId() );
-        assertNotNull( result.getDependantContents().get( 0 ).getPath() );
-        assertNotNull( result.getDependantContents().get( 0 ).getType() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getCompareStatus() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getName() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getDisplayName() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getIconUrl() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getId() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getPath() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getType() );
     }
 
     @Test
@@ -140,17 +117,20 @@ public class ResolveDependantsResultJsonFactoryTest
         final ResolvePublishDependenciesResult resolvedDependencies =
             ResolvePublishDependenciesResult.create().pushContentRequests( pushRequests ).build();
 
-        final ResolveDependantsResultJson result = createDependantsResultJson( resolvedDependencies, resolved, compareResults );
+        final ResolvePublishRequestedContentsResultJson result =
+            createPublishDependenciesResultJson( resolvedDependencies, resolved, compareResults );
 
-        assertEquals( 2, result.getDependantContents().size() );
+        assertEquals( 1, result.getPushRequestedContents().size() );
+        assertEquals( 0, result.getPushRequestedContents().get( 0 ).getChildrenCount() );
+        assertEquals( 2, result.getPushRequestedContents().get( 0 ).getDependantsCount() );
 
-        assertNull( result.getDependantContents().get( 0 ).getCompareStatus() );
-        assertNotNull( result.getDependantContents().get( 0 ).getName() );
-        assertNotNull( result.getDependantContents().get( 0 ).getDisplayName() );
-        assertNotNull( result.getDependantContents().get( 0 ).getIconUrl() );
-        assertNotNull( result.getDependantContents().get( 0 ).getId() );
-        assertNotNull( result.getDependantContents().get( 0 ).getPath() );
-        assertNotNull( result.getDependantContents().get( 0 ).getType() );
+        assertNull( result.getPushRequestedContents().get( 0 ).getCompareStatus() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getName() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getDisplayName() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getIconUrl() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getId() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getPath() );
+        assertNotNull( result.getPushRequestedContents().get( 0 ).getType() );
     }
 
     @Test(expected = IllegalArgumentException.class)
@@ -170,13 +150,13 @@ public class ResolveDependantsResultJsonFactoryTest
         final ResolvePublishDependenciesResult resolvedDependencies =
             ResolvePublishDependenciesResult.create().pushContentRequests( pushRequests ).build();
 
-        createDependantsResultJson( resolvedDependencies, resolved, compareResults );
+        createPublishDependenciesResultJson( resolvedDependencies, resolved, compareResults );
     }
 
-    private ResolveDependantsResultJson createDependantsResultJson( final ResolvePublishDependenciesResult resolvedDependencies,
-                                                                    final Contents resolved, final CompareContentResults compareResults )
+    private ResolvePublishRequestedContentsResultJson createPublishDependenciesResultJson(
+        final ResolvePublishDependenciesResult resolvedDependencies, final Contents resolved, final CompareContentResults compareResults )
     {
-        return ResolveDependantsResultJsonFactory.create().
+        return ResolvePublishRequestedContentsResultJsonFactory.create().
             resolvedPublishDependencies( resolvedDependencies ).
             iconUrlResolver( new ContentIconUrlResolver( this.contentTypeService ) ).
             resolvedContents( resolved ).
