+++ modules/admin-ui/src/main/resources/web/admin/common/js/content/ResolvePublishDependenciesRequest.ts
@@ -25,5 +25,6 @@ module api.content {
         getRequestPath(): api.rest.Path {
             return api.rest.Path.fromParent(super.getResourcePath(), "resolvePublishDependencies");
         }
+
     }
 }
