+++ modules/admin-ui/src/main/resources/web/admin/common/js/content/_module.ts
@@ -61,8 +61,8 @@
 ///<reference path='ContentQueryResult.ts' />
 ///<reference path='UpdateContentRequest.ts' />
 ///<reference path='PublishContentRequest.ts' />
-///<reference path='ResolvePublishDependenciesRequest.ts' />
-///<reference path='ResolveDependantsRequest.ts' />
+///<reference path='ResolvePublishRequestedContentsRequest.ts' />
+///<reference path='resolvePublishDependenciesRequest.ts' />
 ///<reference path='PublishContentResult.ts' />
 ///<reference path='CompareContentRequest.ts' />
 ///<reference path='CompareContentResults.ts' />
