--- modules/admin-ui/src/main/resources/web/admin/common/js/content/json/ResolveDependantsResultJson.ts
@@ -1,7 +1,0 @@
-module api.content.json {
-
-    export interface ResolveDependantsResultJson {
-
-        dependantContents: ResolvedPublishDependencyJson[];
-    }
-}
