+++ modules/admin-ui/src/main/resources/web/admin/common/js/content/json/ResolvePublishDependenciesResultJson.ts
@@ -2,6 +2,8 @@ module api.content.json {
 
     export interface ResolvePublishDependenciesResultJson {
 
-        pushRequestedContents: ResolvedPublishRequestedContentJson[];
+        dependantContents: ResolvedPublishDependencyJson[];
+
+        childrenContents: ResolvedPublishDependencyJson[];
     }
 }
