+++ modules/admin-ui/src/main/resources/web/admin/common/js/content/json/ResolvePublishRequestedContentsResultJson.ts
@@ -1,0 +1,7 @@
+module api.content.json {
+
+    export interface ResolvePublishRequestedContentsResultJson {
+
+        pushRequestedContents: ResolvedPublishRequestedContentJson[];
+    }
+}
